third_party/blink/web_tests/wpt_internal/resource-timing/initiator-type-early-hints-preload.h2.window-expected.txt

This is a testharness.js-based test.
FAIL Ensure initiatorType is set to 'early-hints' assert_equals: expected "early-hints" but got "link"
Harness: the test ran to completion.
